ApplicationStackHelper: relocate to application directory
