+++ xbmc/application/ApplicationStackHelper.cpp
