+++ xbmc/application/ApplicationStackHelper.h
